+++ Editor/lib/Components/components.dart
@@ -7,3 +7,6 @@ abstract class Component {
     throw UnimplementedError();
   }
 }
+
+abstract class IMSComponent{}
+abstract class MSComponent<T extends Component> implements IMSComponent {}